+++ README.md
@@ -18,4 +18,4 @@ it in the initial code):
 
 - Implement a solution following the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline).
 - Open one more terminal and run tests with `npm test` to ensure your solution is correct.
-- Replace `<your_account>` with your Github username in the [DEMO LINK](https://<your_account>.github.io/react_article/) and add it to the PR description.
+- Replace `<your_account>` with your Github username in the [DEMO LINK](https://DenysDubas.github.io/react_article/) and add it to the PR description.
